Add files via upload
Makefile is the same.  Updated to take an IP address for the server to move between client and server on separate machines (#define SRV_IP).  File.txt is literally the main code for replacing files, just used it to test moving files.  Once we have the part that copies files to a single directory and gives a filename array/list/whatever, I can iterate through it instead of doing one file, and that'll be the end of this.  Feel free to take a look or clean as you see fit.